+++ Sources/ZcashLightClientKit/Repository/UnspentTransactionOutputRepository.swift
@@ -10,7 +10,6 @@ import Foundation
 protocol UnspentTransactionOutputRepository {
     func initialise() async throws
     func getAll(address: String?) async throws -> [UnspentTransactionOutputEntity]
-    func balance(address: String, latestHeight: BlockHeight) async throws -> WalletBalance
     func store(utxos: [UnspentTransactionOutputEntity]) async throws
     func clearAll(address: String?) async throws
 }